+++ Samples/SampleSwiftUI/ContentView.swift
@@ -3,52 +3,61 @@ import SwiftUI
 import DevRevSDK
 
 struct ContentView: View {
-	@SwiftUI.State private var isUserConfigured = true
+	@SwiftUI.State private var isConfigured = false
 	@SwiftUI.State private var isUserIdentified = false
-	
+	@SwiftUI.State private var isMonitoringEnabled = false
+
 	var body: some View {
 		NavigationView {
 			List {
-				HStack{
-					Image(systemName: isUserConfigured ? "checkmark.square.fill" : "square")
-					Text("Configured")
+				Section(header: Text("Status")) {
+					StatusRow(
+						status: "Is the SDK configured?",
+						isComplete: $isConfigured
+					)
+					StatusRow(
+						status: "Is the user identified?",
+						isComplete: $isUserIdentified
+					)
+					StatusRow(
+						status: "Is session monitoring enabled?",
+						isComplete: $isMonitoringEnabled
+					)
 				}
-				HStack{
-					Image(systemName: isUserIdentified ? "checkmark.square.fill" : "square")
-					Text("Identified")
-				}
-				NavigationLink(destination: IdentificationView()) {
-					Text("Identification")
-				}
-				NavigationLink(destination: PushNotificationView()) {
-					Text("Push Notification")
-				}
-				NavigationLink(destination: SupportView()) {
-					Text("Support")
-				}
-				.buttonStyle(PlainButtonStyle())
-				NavigationLink(destination: SessionAnalyticsView()) {
-					Text("Session Analytics")
+
+				Section(header: Text("Features")) {
+					NavigationLink(destination: IdentificationView()) {
+						Text("Identification")
+					}
+					NavigationLink(destination: PushNotificationsView()) {
+						Text("Push Notifications")
+					}
+					NavigationLink(destination: SupportView()) {
+						Text("Support")
+					}
+					NavigationLink(destination: SessionAnalyticsView()) {
+						Text("Session Analytics")
+					}
 				}
 			}
-			.navigationTitle("Main Menu")
+			.navigationTitle("DevRev SDK")
+			.navigationBarItems(trailing: RefreshButton(action: updateStatuses))
+			.refreshable {
+				await updateStatuses()
+			}
 			.task {
-				await checkIdentification()
+				await updateStatuses()
 			}
 		}
 	}
-	private func checkIdentification() async
-	{
-		let isUserIdentified = await DevRev.isUserIdentified
-				
-		await MainActor.run {
-			self.isUserIdentified = isUserIdentified
-		}
+
+	private func updateStatuses() async {
+		isConfigured = await DevRev.isConfigured
+		isUserIdentified = await DevRev.isUserIdentified
+		isMonitoringEnabled = DevRev.isMonitoringEnabled
 	}
 }
 
-struct ContentView_Previews: PreviewProvider {
-	static var previews: some View {
-		ContentView()
-	}
+#Preview {
+	ContentView()
 }