+++ Samples/SampleUIKit/Constants/Constants.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+enum Constants {
+	enum CellIdentifier {
+		static let pushNotifications = "PushNotificationsCell"
+		static let identification = "IdentificationCell"
+		static let unverifiedUserIdentification = "UnverifiedUserIdentificationCell"
+		static let verifiedUserIdentification = "VerifiedUserIdentificationCell"
+		static let sessionAnalytics = "SessionAnalyticsCell"
+		static let support = "SupportCell"
+		static let main = "MainCell"
+	}
+}