+++ Samples/SampleUIKit/SceneDelegate.swift
@@ -1,6 +1,25 @@
 import Foundation
 import UIKit
+import DevRevSDK
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 	var window: UIWindow?
+
+	func scene(
+		_ scene: UIScene,
+		willConnectTo session: UISceneSession,
+		options connectionOptions: UIScene.ConnectionOptions
+	) {
+		guard
+			let scene = scene as? UIWindowScene
+		else {
+			return
+		}
+
+		window = .init(windowScene: scene)
+		let rootViewController = ViewController()
+		let navigationController = UINavigationController(rootViewController: rootViewController)
+		window?.rootViewController = navigationController
+		window?.makeKeyAndVisible()
+	}
 }